requirements.txt
flask
certifi==2020.4.5.1
chardet==3.0.4
future==0.18.2
idna==2.9
line-bot-sdk==1.16.0
requests==2.23.0
urllib3==1.25.9